--- jobserver/templates/org_create.html
@@ -1,49 +1,0 @@
-{% extends "base.html" %}
-
-{% block content %}
-
-<div class="row">
-  <div class="col-lg-8 offset-lg-2">
-
-      <h3>Create an Organisation</h3>
-
-      <form method="POST">
-        {% csrf_token %}
-
-        <div class="form-group row">
-          <label for="id_name" class="col-sm-2 col-form-label">Name</label>
-          <div class="col-sm-10">
-            <input class="form-control" id="id_name" name="name" type="text" />
-
-            {% if form.name.errors %}
-            <ul class="pl-3 mb-1">
-              {% for error in form.name.errors %}
-              <li class="text-danger">{{ error }}</li>
-              {% endfor %}
-            </ul>
-            {% endif %}
-
-          </div>
-        </div>
-
-        <button class="btn btn-primary" type="submit">Create</button>
-      </form>
-
-      <h3 class="text-center my-5">OR</h3>
-
-      <h3>Select an existing Organisation</h3>
-
-      <div class="list-group mb-3 overflow-auto list">
-        {% for org in orgs %}
-        <a
-          class="list-group-item list-group-item-action"
-          href="{{ org.get_absolute_url }}">
-          <span class="text-primary">{{ org.name }}</span>
-        </a>
-        {% endfor %}
-      </div>
-
-  </div>
-</div>
-
-{% endblock %}
