+++ jobserver/views/orgs.py
@@ -1,30 +1,9 @@
 from django.http import Http404
 from django.shortcuts import redirect
 from django.template.response import TemplateResponse
-from django.utils.decorators import method_decorator
-from django.views.generic import CreateView, DetailView, ListView
+from django.views.generic import DetailView, ListView
 
-from ..authorization.decorators import require_permission
-from ..forms import OrgCreateForm
 from ..models import Org, Workspace
-
-
-@method_decorator(require_permission("org_create"), name="dispatch")
-class OrgCreate(CreateView):
-    form_class = OrgCreateForm
-    model = Org
-    template_name = "org_create.html"
-
-    def form_valid(self, form):
-        org = form.save(commit=False)
-        org.created_by = self.request.user
-        org.save()
-        return redirect(org)
-
-    def get_context_data(self, **kwargs):
-        context = super().get_context_data(**kwargs)
-        context["orgs"] = Org.objects.order_by("name")
-        return context
 
 
 class OrgDetail(DetailView):
